+++ lib/karafka/pro/base_consumer.rb
@@ -26,11 +26,14 @@ module Karafka
       # Pauses processing of a given partition until we're done with the processing
       # This ensures, that we can easily poll not reaching the `max.poll.interval`
       def on_before_consume
-        # Pause at the first message in a batch. That way in case of a crash, we will not loose
-        # any messages
         return unless topic.long_running_job?
 
-        pause(messages.first.offset, MAX_PAUSE_TIME)
+        # This ensures, that when running LRJ with VP, things operate as expected
+        coordinator.on_started do |first_group_message|
+          # Pause at the first message in a batch. That way in case of a crash, we will not loose
+          # any messages
+          pause(first_group_message.offset, MAX_PAUSE_TIME)
+        end
       end
 
       # Runs extra logic after consumption that is related to handling long running jobs
@@ -39,13 +42,32 @@ module Karafka
         # Nothing to do if we lost the partition
         return if revoked?
 
+        # For virtual partitions we need to make sure we only run the post-consumption code once
+        if topic.virtual_partitioner?
+          # Here we get the first and last messages from all the virtual partitions so we can
+          # threat them conceptually as if they would have been executed not in parallel
+          coordinator.on_finished do |first_group_message, last_group_message|
+            on_after_consume_regular(first_group_message, last_group_message)
+          end
+        else
+          on_after_consume_regular(messages.first, messages.last)
+        end
+      end
+
+      private
+
+      # Handles the post-consumption flow depending on topic settings
+      #
+      # @param first_message [Karafka::Messages::Message]
+      # @param last_message [Karafka::Messages::Message]
+      def on_after_consume_regular(first_message, last_message)
         if coordinator.success?
           coordinator.pause_tracker.reset
 
           # We use the non-blocking one here. If someone needs the blocking one, can implement it
           # with manual offset management
           # Mark as consumed only if manual offset management is not on
-          mark_as_consumed(messages.last) unless topic.manual_offset_management?
+          mark_as_consumed(last_message) unless topic.manual_offset_management?
 
           # We check it twice as marking could change this state
           return if revoked?
@@ -60,12 +82,12 @@ module Karafka
           # interesting (yet valid) corner case, where with manual offset management on and no
           # marking as consumed, we end up with an infinite loop processing same messages over and
           # over again
-          seek(@seek_offset || messages.first.offset)
+          seek(@seek_offset || first_message.offset)
 
           resume
         else
           # If processing failed, we need to pause
-          pause(@seek_offset || messages.first.offset)
+          pause(@seek_offset || first_message.offset)
         end
       end
     end